pyproject.toml
[project]
name = "temperature-adapt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.0",
    "sqlite-utils>=3.36",
    "biopython>=1.80",
]
